+++ sdk/core/src/lib.rs
@@ -21,7 +21,7 @@ mod errors;
 mod http_client;
 mod models;
 mod options;
-mod pipeline;
+mod pageable;
 mod policies;
 mod request;
 mod request_options;
@@ -49,7 +49,7 @@ pub use headers::AddAsHeader;
 pub use http_client::{new_http_client, to_json, HttpClient};
 pub use models::*;
 pub use options::*;
-pub use pipeline::Pipeline;
+pub use pageable::*;
 pub use policies::{Policy, PolicyResult};
 pub use request::*;
 pub use response::*;
